Refactor: Remove deprecated components and consolidate invoice forms
- Deleted ClientInfoFormNew.tsx and InvoiceSummaryNew.tsx as they were backups.
- Removed old LineItemsTable.tsx and corrupted InvoiceSummary.tsx, replacing them with updated versions.
- Cleaned up tailwind.config.ts as it was a duplicate.
