--- tailwind.config.ts
@@ -1,19 +1,0 @@
-import type { Config } from "tailwindcss";
-
-const config: Config = {
-  content: [
-    "./pages/**/*.{js,ts,jsx,tsx,mdx}",
-    "./components/**/*.{js,ts,jsx,tsx,mdx}",
-    "./app/**/*.{js,ts,jsx,tsx,mdx}",
-  ],
-  theme: {
-    extend: {
-      fontFamily: {
-        inter: ['var(--font-inter)', 'sans-serif'],
-      },
-    },
-  },
-  plugins: [require("daisyui")],
-};
-
-export default config;
